+++ src/main/java/de/mainzelliste/paths/webservice/PathsResource.java
@@ -9,8 +9,10 @@ import javax.ws.rs.OPTIONS;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
@@ -21,6 +23,7 @@ public class PathsResource {
 
     @POST
     @Path("/{pathName}")
+    @Produces(MediaType.APPLICATION_JSON)
     public Response executePath(@Context HttpServletRequest req, @PathParam("pathName") String pathName, String data) {
         Controller.authenticator.checkPermission(req, pathName);
         AbstractProcessor implementation = backend.getPathImplementation(pathName);